+++ src/rovr/action_buttons/rename_item_button.py
@@ -90,7 +90,7 @@ class RenameItemButton(Button):
                     severity="error",
                 )
         try:
-            self.app.file_list_pause_check = True  # ty: ignore[invalid-assignment]
+            self.app.file_list.file_list_pause_check = True  # ty: ignore[invalid-assignment]
             self.app.file_list.focus()
             worker: Worker = self.app.file_list.update_file_list(
                 add_to_session=False, focus_on=path.basename(new_name)
@@ -98,4 +98,4 @@ class RenameItemButton(Button):
             with contextlib.suppress(WorkerError):
                 await worker.wait()
         finally:
-            self.app.file_list_pause_check = False  # ty: ignore[invalid-assignment]
+            self.app.file_list.file_list_pause_check = False  # ty: ignore[invalid-assignment]