+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## new features
 - cli: add `--config-folder` option to specify custom config folder #185
 - preview: batched pdf loading #184
+- filelist: dim files/folders that are cut in clipboard #188
 
 <details>
 <summary><h1>[v0.7.0.dev1]</h1></summary>